+++ catalog/app/containers/Bucket/Queries/Athena/CreatePackage.tsx
@@ -24,25 +24,25 @@ function SeeDocsForCreatingPackage() {
 }
 
 function doQueryResultsContainManifestEntries(
-  rows: string[][],
-): rows is [ManifestKey[], ...string[][]] {
-  const [head] = rows
+  queryResults: requests.athena.QueryResultsResponse,
+): queryResults is requests.athena.QueryManifestsResponse {
+  const columnNames = queryResults.columns.map(({ name }) => name)
   return (
-    head.includes('size') &&
-    head.includes('physical_keys') &&
-    head.includes('logical_key')
+    columnNames.includes('size') &&
+    columnNames.includes('physical_keys') &&
+    columnNames.includes('logical_key')
   )
 }
 
 function rowToManifestEntryStringified(
   row: string[],
-  head: ManifestKey[],
+  columns: requests.athena.QueryResultsColumns,
 ): ManifestEntryStringified {
   return row.reduce((acc, value, index) => {
-    if (!head[index]) return acc
+    if (!columns[index].name) return acc
     return {
       ...acc,
-      [head[index]]: value,
+      [columns[index].name]: value,
     }
   }, {} as ManifestEntryStringified)
 }
@@ -72,11 +72,10 @@ function parseManifestEntryStringified(entry: ManifestEntryStringified): {
 }
 
 function parseQueryResults(
-  rows: [ManifestKey[], ...string[][]],
+  queryResults: requests.athena.QueryManifestsResponse,
 ): Record<string, Model.S3File> {
-  const [head, ...tail] = rows
-  const manifestEntries: ManifestEntryStringified[] = tail.reduce(
-    (memo, row) => memo.concat(rowToManifestEntryStringified(row, head)),
+  const manifestEntries: ManifestEntryStringified[] = queryResults.rows.reduce(
+    (memo, row) => memo.concat(rowToManifestEntryStringified(row, queryResults.columns)),
     [] as ManifestEntryStringified[],
   )
   return manifestEntries.reduce(
@@ -90,10 +89,10 @@ function parseQueryResults(
 
 interface CreatePackageProps {
   bucket: string
-  rows: requests.athena.QueryResultsRows
+  queryResults: requests.athena.QueryResultsResponse
 }
 
-export default function CreatePackage({ bucket, rows }: CreatePackageProps) {
+export default function CreatePackage({ bucket, queryResults }: CreatePackageProps) {
   const addToPackage = AddToPackage.use()
   const createDialog = usePackageCreationDialog({
     bucket,
@@ -101,15 +100,17 @@ export default function CreatePackage({ bucket, rows }: CreatePackageProps) {
     disableStateDisplay: true,
   })
   const onPackage = React.useCallback(() => {
-    if (!doQueryResultsContainManifestEntries(rows)) return
+    if (!doQueryResultsContainManifestEntries(queryResults)) return
 
     // TODO: make it lazy, and disable button
-    const entries = parseQueryResults(rows)
+    const entries = parseQueryResults(queryResults)
     addToPackage?.merge(entries)
     createDialog.open()
-  }, [addToPackage, createDialog, rows])
+  }, [addToPackage, createDialog, queryResults])
 
-  if (!doQueryResultsContainManifestEntries(rows)) return <SeeDocsForCreatingPackage />
+  if (!doQueryResultsContainManifestEntries(queryResults)) {
+    return <SeeDocsForCreatingPackage />
+  }
 
   return (
     <>